flowmanager/requirements.txt
APScheduler==3.7.0
asgiref==3.2.3
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
pytz==2019.3
six==1.16.0
sqlparse==0.3.1
tzlocal==2.1
